feat: flask delivering static generated by React
