+++ sample-apps/11-client/package.json
@@ -1,22 +1,18 @@
 {
-  "name": "observability-with-amazon-opensearch-webapp-client",
+  "name": "observability-with-amazon-opensearch-react-client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@cloudscape-design/components": "^3.0.39",
+    "@cloudscape-design/components": "^3.0.60",
     "@cloudscape-design/global-styles": "^1.0.1",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
     "@types/node": "^16.11.56",
     "@types/react": "^18.0.18",
     "@types/react-dom": "^18.0.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-scripts": "^5.0.1",
     "typescript": "^4.8.2",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^3.0.2"
   },
   "scripts": {
     "start": "react-scripts start -p 5005",
@@ -28,8 +24,7 @@
   "proxy": "http://localhost:5000",
   "eslintConfig": {
     "extends": [
-      "react-app",
-      "react-app/jest"
+      "react-app"
     ]
   },
   "browserslist": {
