+++ sample-apps/11-client/src/index.tsx
@@ -5,7 +5,7 @@ import App from './App';
 import reportWebVitals from './reportWebVitals';
 
 import "@cloudscape-design/global-styles/index.css"
-import "index.css"
+import "./index.css"
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
