+++ sample-apps/11-client/tsconfig.json
@@ -18,13 +18,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx",
-    "baseUrl": "./src",
-    "paths": {
-      "@components/*": [
-        "./components/*"
-      ]
-    }
+    "jsx": "preserve"
   },
   "include": [
     "src"
