--- sample-apps/13-client/Error.py
@@ -1,16 +1,0 @@
-from flask import jsonify
-
-class Error(Exception):
-    status_code = 400
-
-    def __init__(self, message, status_code=None, payload=None):
-        Exception.__init__(self)
-        self.message = message
-        if status_code is not None:
-            self.status_code = status_code
-        self.payload = payload
-
-    def to_dict(self):
-        result = dict(self.payload or ())
-        result['message'] = self.message
-        return result
