--- sample-apps/13-client/kubernetes/00-namespace.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: client-service
-
