--- sample-apps/13-client/kubernetes/01-deployment.yaml
@@ -1,62 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    app: client-service
-  name: client-service
-  namespace: client-service
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: client-service
-  template:
-    metadata:
-      labels:
-        app: client-service
-    spec:
-      containers:
-      - image: __ACCOUNT_ID__.dkr.ecr.__AWS_REGION__.amazonaws.com/client-service
-        name: client-service
-        ports:
-        - containerPort: 5000
-        resources: {} # Verify the amount of requests
-          # requests:
-          #   memory: "64Mi"
-          #   cpu: "250m"
-          # limits:
-          #   memory: "128Mi"
-          #   cpu: "500m"
-        env:
-        - name: OTLP
-          value: "otel-collector.otel-collector.svc.cluster.local"
-        - name: RECOMMEND
-          value: "recommendation-service.recommendation-service.svc.cluster.local"
-        - name: LOGS
-          value: "analytics-service.analytics-service.svc.cluster.local" # Analytics Service
-        - name: ORDER
-          value: "order-service.order-service.svc.cluster.local"
-        - name: INVENTORY
-          value: "inventory-service.inventory-service.svc.cluster.local"
-        - name: PAYMENT
-          value: "payment-service.payment-service.svc.cluster.local"
-        - name: AUTH
-          value: "authentication-service.authentication-service.svc.cluster.local"
-        - name: MYSQL_HOST
-          value: "mysql.mysql.svc.cluster.local"
-        - name: MYSQL_PORT
-          value: "3306"
-        
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: client-service
-  namespace: client-service
-spec:
-  type: LoadBalancer
-  ports:
-  - port: 80
-    targetPort: 5000
-  selector:
-    app: client-service
